Adicionando protecao nas rotas com jwt

+++ src/main/java/unoeste/fipp/mercadofipp/restcontrollers/CategoriaRestController.java
@@ -1,5 +1,6 @@
 package unoeste.fipp.mercadofipp.restcontrollers;
 
+import io.jsonwebtoken.Claims;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,18 +18,20 @@ public class CategoriaRestController {
     @Autowired
     private CategoriaService categoriaService;
 
-    @GetMapping
-    public ResponseEntity<Object> getAll(@RequestHeader("Authorization") String token){
-        if (token != null && token.startsWith("Bearer ")) {
+    public boolean verificaPermissao(String token) {
+        if(token!= null && token.startsWith("Bearer "))
             token = token.substring(7);
-        }
-        System.out.println("CoNTROLER: " + token);
-        if (!JWTTokenProvider.verifyToken(token)) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-                    .body(new Erro("Token inválido ou expirado"));
-        }
-        List<Categoria> categoriaList= categoriaService.getAll();
 
+        Claims dados = JWTTokenProvider.getAllClaimsFromToken(token);
+        String level = dados.get("level", String.class);
+        if(!level.equals("1"))
+            return false;
+        return true;
+    }
+
+    @GetMapping
+    public ResponseEntity<Object> getAll(){
+        List<Categoria> categoriaList= categoriaService.getAll();
 
         if(categoriaList!=null && !categoriaList.isEmpty()){
             return ResponseEntity.ok(categoriaList);
@@ -37,7 +40,9 @@ public class CategoriaRestController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Object> getById(@PathVariable(name = "id")long id){
+    public ResponseEntity<Object> getById(@PathVariable(name = "id")long id, @RequestHeader("Authorization") String token){
+        if(!verificaPermissao(token))
+            return  ResponseEntity.badRequest().body(new Erro("Usuario sem permissão"));
        Categoria categoria= categoriaService.getById(id);
         if(categoria!=null){
             return ResponseEntity.ok(categoria);
@@ -46,21 +51,28 @@ public class CategoriaRestController {
     }
 
     @PostMapping
-    public ResponseEntity<Object> inserir(@RequestBody Categoria categoria){
+    public ResponseEntity<Object> inserir(@RequestBody Categoria categoria, @RequestHeader("Authorization") String token){
+
+        if(!verificaPermissao(token))
+            return  ResponseEntity.badRequest().body(new Erro("Usuario sem permissão"));
         if(categoriaService.save(categoria)!=null)
             return ResponseEntity.ok(categoria);
         return ResponseEntity.badRequest().body(new Erro("Erro ao cadastrar uma nova categoria"));
     }
 
     @PutMapping
-    public ResponseEntity<Object> alterar(@RequestBody Categoria categoria){
+    public ResponseEntity<Object> alterar(@RequestBody Categoria categoria, @RequestHeader("Authorization") String token){
+        if(!verificaPermissao(token))
+            return  ResponseEntity.badRequest().body(new Erro("Usuario sem permissão"));
         if(categoriaService.save(categoria)!=null)
             return ResponseEntity.ok(categoria);
         return ResponseEntity.badRequest().body(new Erro("Erro ao alterar uma nova categoria"));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Object> deletar(@PathVariable long id){
+    public ResponseEntity<Object> deletar(@PathVariable long id, @RequestHeader("Authorization") String token){
+        if(!verificaPermissao(token))
+            return  ResponseEntity.badRequest().body(new Erro("Usuario sem permissão"));
         if(categoriaService.delete(id))
             return ResponseEntity.noContent().build();
         return ResponseEntity.badRequest().body("Erro ao deletar categoria!");
